requirements.txt
aiogram>=3.0.0
gspread
oauth2client
tenacity
python-dotenv
